+++ contracts/index.ts
@@ -1,5 +1,6 @@
 export * from "./analyticsEvents";
 export * from "./gameCatalog";
+export * from "./moderationSafety";
 export * from "./progression";
 export * from "./roomMatch";
 export * from "./userProfile";